Extract Contact superclass from Phone, Email and Address
