+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,23 @@
+package seedu.addressbook.data.person;
+
+public abstract class Contact {
+    public String value;
+    protected boolean isPrivate;
+
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+
+
+}
